requirements.txt
Flask
python-dotenv
pymongo
openai
pandas
flask-cors
PyJWT